last

+++ app/controllers/bucket_controller.rb
@@ -1,6 +1,6 @@
 class BucketController < ApplicationController
 
-    before_action :authenticate_user!, only:[:make_bucket, :support]
+    before_action :authenticate_user!, only:[:write_comment, :del_comment, :make_bucket, :create_bucket, :modify ,:modify_bucket, :delete_bucket, :support, :support_send]
 
     def projectList
         @first_bucket_count=Bucket.where(:state => 0).count
@@ -16,28 +16,16 @@ class BucketController < ApplicationController
 
     def secondProcessList
         @second_buckets=Bucket.where(:state => 1)
-        if @second_buckets.count == 0
-            flash[:danger] = "2단계 버킷이 없습니다."
-            redirect_to "/bucket/firstProcessList"
-        else
-        end
     end
     
 
     def finalProcessList
         @final_buckets=Bucket.where(:state => 2)
-        if @final_buckets.count == 0
-            flash[:danger] = "3단계 버킷이 없습니다."
-            redirect_to "/bucket/firstProcessList"
-        else
-        end
     end
 
 
     def make_bucket
-
-        @name =  params[:main_title_input]
-
+        @title=params[:main_title_input]
     end	
 
 
@@ -57,10 +45,8 @@ class BucketController < ApplicationController
         redirect_to :back
     end
 
-    def new
-        @bucket = Bucket.new
-    end
-    
+
+
     def create_bucket
 
         item_count=params[:item_count]
@@ -75,6 +61,12 @@ class BucketController < ApplicationController
           bucket.thumbnail = params[:thumbnail]
         end
 
+        ## fastImage
+        # if (params[:thumbnail]!=nil)
+        #     outfile = FastImage.resize(params[:thumbnail], 500, 500)            
+        #     bucket.thumbnail = outfile
+        # end
+        
         bucket.start_date=convert_date(params[:start_date])
         bucket.finish_date=convert_date(params[:finish_date])
         bucket.state=0
@@ -98,41 +90,44 @@ class BucketController < ApplicationController
             flash[:danger] = "저장에 실패하였습니다."
             redirect_to :back
         end
+
     end
 
-    def create_bucket2
-        @bucket=Bucket.new(bucket_params)
-        @bucket.state = 0
-        if @bucket.save
-            flash[:success] = "정상적으로 저장되었습니다."
-            redirect_to "/"
-        else 
-            session[:error] = @bucket.errors.full_messages
-            redirect_to :back
-        end
-    end 
-    private
-        def bucket_params
-            params.require(:bucket).permit(:name, :intro_simple, :intro_detail, :start_date, :final_date)
-        end
 
-    public
+    # def create_bucket
+    #     @bucket=Bucket.new(bucket_params)
+    #     @bucket.state = 0
+    #     if @bucket.save
+    #         flash[:success] = "정상적으로 저장되었습니다."
+    #         redirect_to "/"
+    #     else 
+    #         session[:error] = @bucket.errors.full_messages
+    #         redirect_to :back
+    #     end
+    # end 
+
+    # private
+    #     def bucket_params
+    #         params.require(:bucket).permit(:name, :intro_simple, :intro_detail, :start_date, :final_date)
+    #     end
+
+    # public
+
     def convert_date (arg0)
-        if arg0 == ""
-        else
-            date=""
-            date_array=arg0.split('/')
-            date_array.insert(0,date_array.pop)
 
-            date_array.each do |x|
-                date << x
-                if date.length!=10
-                    date << '-'
-                end
+        date=""
+        date_array=arg0.split('/')
+        date_array.insert(0,date_array.pop)
+
+        date_array.each do |x|
+            date << x
+            if date.length!=10
+                date << '-'
             end
-
-            return date
         end
+
+        return date
+
     end
 
     def convert_date_reverse (arg0)
@@ -171,45 +166,63 @@ class BucketController < ApplicationController
         bucket.intro_simple=params[:intro_simple]
         bucket.intro_detail=params[:intro_detail]
 
-        if (params[:thumbnail]!=nil)
-         bucket.thumbnail = params[:thumbnail]
-        end
+        #if (params[:thumbnail]!=nil)
+        #  bucket.thumbnail = params[:thumbnail]
+        #end
 
-        # if (params[:thumbnail]!=nil)
-        #     outfile = FastImage.resize(params[:thumbnail], 500, 500)            
-        #     bucket.thumbnail = outfile
-        # end
+        if (params[:thumbnail]!=nil)
+            outfile = FastImage.resize(params[:thumbnail], 500, 500)            
+            bucket.thumbnail = outfile
+        end
         
         bucket.start_date=convert_date(params[:start_date])
         bucket.finish_date=convert_date(params[:finish_date])
 
-        bucket.save
+        if bucket.save
 
 
-        bucket.items.where(:state => 0).each do |x|
+            bucket.items.where(:state => 0).each do |x|
+                x.destroy
+            end
 
-            x.destroy
+            1.upto(item_count.to_i) do |i|
 
+                item=Item.new
+                item.bucket_id=bucket.id
+                item.name=params['item_'+i.to_s]
+                item.intro=params['item_'+i.to_s+'_intro']
+                item.state=0
+                item.save
+                # 빈칸 걸러내야 함
+            end
+            flash[:success] =  "성공적으로 저장되었습니다."
+            redirect_to "/bucket/projectdetail/#{bucket.id}"
+        else
+            flash[:danger] = "저장에 실패하였습니다."
+            redirect_to :back
         end
-
-        1.upto(item_count.to_i) do |i|
-
-            item=Item.new
-            item.bucket_id=bucket.id
-            item.name=params['item_'+i.to_s]
-            item.intro=params['item_'+i.to_s+'_intro']
-            item.state=0
-            item.save
-            # 빈칸 걸러내야 함
-        end
-
-        redirect_to "/bucket/projectdetail/#{bucket.id}"
-
 
     end
 
     def delete_bucket
         one_bucket = Bucket.find(params[:id])
+
+        all_items = one_bucket.items
+
+        all_items.each do |i|
+            i.destroy
+        end
+
+        one_bucket.replies.each do |r|
+            r.destroy
+        end
+
+
+        one_bucket.supports.each do |s|
+            s.destroy
+        end
+
+
         #state값 받아서 redirect할때 state 값을 기준으로 다른 페이지로 보내자
         if current_user.id == one_bucket.user_id
             one_bucket.destroy
@@ -224,9 +237,36 @@ class BucketController < ApplicationController
 
 
     def support
+        @selected_bucket=Bucket.find(params[:id]) # 사용자 정보와 등록버킷리스트 간단히 보여줌.
+    
+    end
+    
+    def support_send
+        @selected_bucket=Bucket.find(params[:bucket_id])
+        itemlist = Array.new
+
+        params[:item_id].each do |i|
+
+            support=Support.new
+            support.user_id=current_user.id
+            support.bucket_id=params[:id]
+
+            item=Item.find(i.to_i)
+            itemlist  << item.name
+            item.state=1
+            item.save
+
+            support.item_id=item.id
+            support.save
+        end
 
 
+
+        UserMail.support_notice(@selected_bucket.user,current_user,itemlist).deliver_now
+       
+        flash[:success]  = "정상적으로 후원이 완료되었습니다. <br>"+@selected_bucket.user.name+"님의 이메일로 fuller"+current_user.name+"님의 메일주소가 전송 되었습니다."
+        redirect_to "/bucket/projectdetail/#{@selected_bucket.id}"
+        
     end
 
-    
 end
